Se agrega domicilios y parejas

+++ app/Http/Controllers/ParticipacionEnEmpresasSociedadesYAsociacionesController.php
@@ -4,11 +4,16 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Arr;
+use App\Declaracion;
+use App\motivoBaja;
 
 class ParticipacionEnEmpresasSociedadesYAsociacionesController extends Controller {
 
     private $request;
+
     public function __construct(Request $request) {
+        $this->middleware("auth");
+        $this->middleware('CheckDeclaracion');
         $this->request = $request;
     }
 
@@ -17,13 +22,11 @@ class ParticipacionEnEmpresasSociedadesYAsociacionesController extends Controlle
      *
      * @return \Illuminate\Http\Response
      */
-
-
-    public function index()
-    {
+    public function index() {
         $empresas = \App\ParticipacionEmpresa::where('declaracion_id', $this->request->session()->get('declaracion_id'))->get();
-        return view("ParticipacionEmpresas.index", compact('empresas'));
-
+        $motivos = Arr::pluck(motivoBaja::all(), "valor", "id");
+        
+        return view("ParticipacionEmpresas.index", compact('empresas', 'motivos'));
     }
 
     /**
@@ -31,8 +34,7 @@ class ParticipacionEnEmpresasSociedadesYAsociacionesController extends Controlle
      *
      * @return \Illuminate\Http\Response
      */
-    public function create()
-    {
+    public function create() {
         //COMBO TITULAR PARTICIPACION
         $RelacionTransmisor = Arr::pluck(\App\RelacionTransmisor::all(), "valor", "id");
 
@@ -49,21 +51,21 @@ class ParticipacionEnEmpresasSociedadesYAsociacionesController extends Controlle
         $selectubicacionParticipacion = Arr::pluck(\App\LugarUbicacion::all(), "valor", "id");
 
         //COMBO TIPO SECTOR
-        $selectsectorProductivo= Arr::pluck(\App\sector::all(), "valor", "id");
+        $selectsectorProductivo = Arr::pluck(\App\sector::all(), "valor", "id");
 
         //COMBO PAIS
-        $selectpais= Arr::pluck(\App\Pais::all(), "valor", "id");
+        $selectpais = Arr::pluck(\App\Pais::all(), "valor", "id");
 
         //COMBO PAIS
-        $selectEntidad= Arr::pluck(\App\Entidad::all(), "entidad", "id");
+        $selectEntidad = Arr::pluck(\App\Entidad::all(), "entidad", "id");
 
-        $tipoSociedad= Arr::pluck(\App\tipoSociedad::all(), "valor", "id");
+        $tipoSociedad = Arr::pluck(\App\tipoSociedad::all(), "valor", "id");
 
-        $tipoModalidad= Arr::pluck(\App\tipoModalidad::all(), "valor", "id");
+        $tipoModalidad = Arr::pluck(\App\tipoModalidad::all(), "valor", "id");
 
         $tipoOperacion = 1;
 
-        return view("ParticipacionEmpresas.create", compact('selecttitularParticipacion', 'selecttipoParticipacion', 'selecttipoRespuesta', 'selectubicacionParticipacion', 'selectsectorProductivo', 'selectpais', 'selectEntidad','RelacionTransmisor','tipoSociedad','tipoModalidad','tipoOperacion'));
+        return view("ParticipacionEmpresas.create", compact('selecttitularParticipacion', 'selecttipoParticipacion', 'selecttipoRespuesta', 'selectubicacionParticipacion', 'selectsectorProductivo', 'selectpais', 'selectEntidad', 'RelacionTransmisor', 'tipoSociedad', 'tipoModalidad', 'tipoOperacion'));
     }
 
     /**
@@ -72,24 +74,25 @@ class ParticipacionEnEmpresasSociedadesYAsociacionesController extends Controlle
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
-    {
+    public function store(Request $request) {
         $empresas = $request->input('empresas');
 //        dd($empresas);
-        $empresas['declaracion_id'] = $request -> session() -> get('declaracion_id');
-        if($empresas["pais_id"] == 0){
+        $empresas['declaracion_id'] = $request->session()->get('declaracion_id');
+        if ($empresas["pais_id"] == 0) {
             $empresas["pais_id"] = null;
         }
+        
+        $empresas["tipo_operaciones_id"] = 1;
+        
         $part_emp = \App\ParticipacionEmpresa::create($empresas);
         return redirect()->route('participacion_empresas.index');
 
-       // dd($empresas);
+        // dd($empresas);
         /*
-        $inmueble->domicilio()->create($domicilio);
-        return redirect()->route('bienes_inmuebles.index');
+          $inmueble->domicilio()->create($domicilio);
+          return redirect()->route('bienes_inmuebles.index');
 
          */
-
     }
 
     /**
@@ -98,8 +101,7 @@ class ParticipacionEnEmpresasSociedadesYAsociacionesController extends Controlle
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
-    {
+    public function show($id) {
         //
     }
 
@@ -109,8 +111,7 @@ class ParticipacionEnEmpresasSociedadesYAsociacionesController extends Controlle
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
-    {
+    public function edit($id) {
         //COMBO TITULAR PARTICIPACION
         $RelacionTransmisor = Arr::pluck(\App\RelacionTransmisor::all(), "valor", "id");
 
@@ -127,23 +128,23 @@ class ParticipacionEnEmpresasSociedadesYAsociacionesController extends Controlle
         $selectubicacionParticipacion = Arr::pluck(\App\LugarUbicacion::all(), "valor", "id");
 
         //COMBO TIPO SECTOR
-        $selectsectorProductivo= Arr::pluck(\App\sector::all(), "valor", "id");
+        $selectsectorProductivo = Arr::pluck(\App\sector::all(), "valor", "id");
 
         //COMBO PAIS
-        $selectpais= Arr::pluck(\App\Pais::all(), "valor", "id");
+        $selectpais = Arr::pluck(\App\Pais::all(), "valor", "id");
 
         //COMBO PAIS
-        $selectEntidad= Arr::pluck(\App\Entidad::all(), "entidad", "id");
+        $selectEntidad = Arr::pluck(\App\Entidad::all(), "entidad", "id");
 
-        $tipoSociedad= Arr::pluck(\App\tipoSociedad::all(), "valor", "id");
+        $tipoSociedad = Arr::pluck(\App\tipoSociedad::all(), "valor", "id");
 
-        $tipoModalidad= Arr::pluck(\App\tipoModalidad::all(), "valor", "id");
+        $tipoModalidad = Arr::pluck(\App\tipoModalidad::all(), "valor", "id");
 
         $empresas = \App\ParticipacionEmpresa::find($id);
 
         $tipoOperacion = 1;
 
-        return view("ParticipacionEmpresas.edit", compact('selecttitularParticipacion', 'selecttipoParticipacion', 'selecttipoRespuesta', 'selectubicacionParticipacion', 'selectsectorProductivo', 'selectpais', 'selectEntidad','RelacionTransmisor','tipoModalidad','tipoSociedad','empresas','tipoOperacion'));
+        return view("ParticipacionEmpresas.edit", compact('selecttitularParticipacion', 'selecttipoParticipacion', 'selecttipoRespuesta', 'selectubicacionParticipacion', 'selectsectorProductivo', 'selectpais', 'selectEntidad', 'RelacionTransmisor', 'tipoModalidad', 'tipoSociedad', 'empresas', 'tipoOperacion'));
     }
 
     /**
@@ -153,20 +154,23 @@ class ParticipacionEnEmpresasSociedadesYAsociacionesController extends Controlle
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
-    {
+    public function update(Request $request, $id) {
         $empresaRow = \App\ParticipacionEmpresa::find($id);
 
         $empresas = $request->input('empresas');
 //        dd($empresas);
-        $empresas['declaracion_id'] = $request -> session() -> get('declaracion_id');
-        if($empresas["pais_id"] == 0){
+        $empresas['declaracion_id'] = $request->session()->get('declaracion_id');
+        if ($empresas["pais_id"] == 0) {
             $empresas["pais_id"] = null;
         }
-
+        
+        if ($empresaRow->enviado) {
+            $empresas["tipo_operaciones_id"] = 2;
+        }
+        
+        
         $empresaRow->update($empresas);
         return redirect()->route('participacion_empresas.index');
-
     }
 
     /**
@@ -175,10 +179,16 @@ class ParticipacionEnEmpresasSociedadesYAsociacionesController extends Controlle
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
-    {
-        $empresa = \App\ParticipacionEmpresa::find($id);
-        $empresa->delete();
+    public function destroy($id) {
+        $empresa = \App\ParticipacionEmpresa::find($this->request->id);
+        
+        if (!$empresa->enviado) {
+            $empresa->delete();
+        } else {
+            $empresa->update(["tipo_operaciones_id" => 4, "motivo_baja_id" => $this->request->motivo_baja_id]);
+        }
+        
         return redirect()->back();
     }
+
 }
